+++ tableaudocumentapi/connection.py
@@ -3,6 +3,8 @@
 # Connection - A class for writing connections to Tableau files
 #
 ###############################################################################
+import xml.etree.ElementTree as ET
+from tableaudocumentapi.dbclass import is_valid_dbclass
 
 
 class Connection(object):
@@ -31,6 +33,17 @@ class Connection(object):
 
     def __repr__(self):
         return "'<Connection server='{}' dbname='{}' @ {}>'".format(self._server, self._dbname, hex(id(self)))
+
+    @classmethod
+    def from_attributes(cls, server, dbname, username, dbclass, authentication=''):
+        root = ET.Element('connection', authentication=authentication)
+        xml = cls(root)
+        xml.server = server
+        xml.dbname = dbname
+        xml.username = username
+        xml.dbclass = dbclass
+
+        return xml
 
     ###########
     # dbname
@@ -111,3 +124,12 @@ class Connection(object):
     @property
     def dbclass(self):
         return self._class
+
+    @dbclass.setter
+    def dbclass(self, value):
+
+        if not is_valid_dbclass(value):
+            raise AttributeError("'{}' is not a valid database type".format(value))
+
+        self._class = value
+        self._connectionXML.set('class', value)