+++ test/test_multidict.py
@@ -22,6 +22,10 @@ class MLDTests(unittest.TestCase):
             }
         })
 
+    def test_multilookupdict_can_be_empty(self):
+        mld = MultiLookupDict()
+        self.assertIsNotNone(mld)
+
     def test_multilookupdict_name_only(self):
         actual = self.mld['[baz]']
         self.assertEqual(3, actual['value'])
@@ -61,3 +65,16 @@ class MLDTests(unittest.TestCase):
     def test_multilookupdict_get_returns_value(self):
         actual = self.mld.get('baz')
         self.assertEqual(1, actual['value'])
+
+    def test_multilookupdict_can_set_item(self):
+        before = self.mld['baz']
+        self.mld['baz'] = 4
+        self.assertEqual(4, self.mld['baz'])
+
+    def test_multilookupdict_can_set_new_item(self):
+        self.mld['wakka'] = 1
+        self.assertEqual(1, self.mld['wakka'])
+
+    def test_multilookupdict_can_set_with_alias(self):
+        self.mld['bar'] = 2
+        self.assertEqual(2, self.mld['[foo]'])